refactor: inference_time_scaling -> its_hub
