+++ its_hub/algorithms/self_consistency.py
