+++ pyproject.toml
@@ -3,7 +3,7 @@ requires = ["setuptools>=42", "wheel"]
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "inference_time_scaling"
+name = "its_hub"
 version = "0.1.0"
 description = "A Python library for inference-time scaling LLMs"
 authors = [
@@ -23,7 +23,7 @@ dependencies = [
 ]
 
 [project.urls]
-Homepage = "https://github.com/Red-Hat-AI-Innovation-Team/inference_time_scaling"
+Homepage = "https://github.com/Red-Hat-AI-Innovation-Team/its_hub"
 
 [project.optional-dependencies]
 dev = [
@@ -31,7 +31,7 @@ dev = [
 ]
 
 [tool.setuptools]
-packages = ["inference_time_scaling"]
+packages = ["its_hub"]
 
 [tool.black]
 line-length = 88
